Permite cantidades con decimales (peso/volumen) en stock, ingresos, salidas, compras y traslados
Antes las columnas de cantidad eran INTEGER, forzando a redondear
cualquier producto medido en kg/litros. Cambia a DECIMAL(10,2):
productos.stock/stock_minimo, venta_detalles, ingreso_detalles,
salida_detalles, orden_compra_detalles y public.orden_traslado_detalles
(cantidad, stock_origen_snapshot, stock_destino_snapshot).

Actualiza el backend (almacen/api.php, traslados/api.php) de
intval()/(int) a floatval()/(float) para las cantidades, y el
frontend (almacen/index.php) para permitir decimales en los inputs
de Ingresos, Salidas y Traslados (quita el step=1 implícito).

migration_35_cantidades_decimales.sql aplicada en local; falta
correr en el VPS.

+++ database/schema_public.sql
@@ -255,10 +255,10 @@ CREATE TABLE IF NOT EXISTS public.orden_traslado_detalles (
     traslado_id             INTEGER NOT NULL REFERENCES public.ordenes_traslado(id) ON DELETE CASCADE,
     producto_codigo         VARCHAR(50) NOT NULL,
     producto_nombre         VARCHAR(200) NOT NULL,
-    cantidad                INTEGER NOT NULL,
+    cantidad                DECIMAL(10,2) NOT NULL,
     costo_unitario          DECIMAL(10,2) DEFAULT 0,
-    stock_origen_snapshot   INTEGER DEFAULT 0,
-    stock_destino_snapshot  INTEGER DEFAULT 0,
+    stock_origen_snapshot   DECIMAL(10,2) DEFAULT 0,
+    stock_destino_snapshot  DECIMAL(10,2) DEFAULT 0,
     observaciones           TEXT,
     created_at              TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
     CONSTRAINT chk_orden_traslado_detalles_cantidad
